HSEARCH-1401 Move basic ORM metadata collection to a dedicated type
We will need to make this more complex for dynamic-map support,
so we'd better separate it from the bootstrap introspector.
